Cast cookie lifetime to int

+++ src/Component/Config/ConfigProvider.php
@@ -25,7 +25,7 @@ final class ConfigProvider implements ConfigProviderInterface
 
     public function getTokenLifetime(): int
     {
-        return $_ENV['TOKEN_LIFETIME'] ?? 1086400;
+        return (int) ($_ENV['TOKEN_LIFETIME'] ?? 1086400);
     }
 
     public function getLogLevel(): string
